Small optimisation to Columns subset serialization
It is possible to disambiguate the small serialization
format from the large with fewer bytes. Prior to this patch
we serialized -1L for the small coding bitmap, however this
value can instead by the column size delta, and it is entirely
consistent: zero means no columns missing in both formats;
if any are missing, we use the size of the superset to decide.