Adds "shadowable" row tombstones for MV timestamp issues
The patch allow row tombstones to be either regular (as before this
patch) or shadowable. If a row tombstone is shadowable, it only applies to the row
and its content as long as the row timestamp is less than the tombstone
timestamp. As soon as the row is updated with a timestamp greater than
the one of the shadowable tombstone, that tombstone is removed, even if some
cells inside the row may have older timestamps.

Patch by Sylvain Lebresne and tjake; reviewed by Tyler Hobbs for CASSANDRA-10261