Accept empty selections in ColumnFilter builder
patch by slebresne; reviewed by aweisberg for CASSANDRA-10471

The builder for ColumnFilter was asserting that the built selection was
at least selection one column. But some empty IN queries actually
select nothing and so that assertion was triggered on some tests.
The patch modify the builder so it accepts that case and return an
empty filter as expected.